Updated types
